Feat. move updatingField logic into proper FieldList component
